refactor: 중간점검 보완
