+++ package.json
@@ -34,6 +34,7 @@
     "react-dom": "19.1.0",
     "react-hook-form": "^7.62.0",
     "react-indiana-drag-scroll": "^2.2.1",
+    "swiper": "^11.2.10",
     "tailwind-merge": "^3.3.1",
     "zod": "^4.0.17",
     "zustand": "^5.0.8"
