+++ src/app/globals.css
@@ -243,10 +243,7 @@ body {
     @apply z-[999];
   }
   .z-header {
-    @apply z-[99];
-  }
-  .z-header-float {
-    @apply z-[100];
+    @apply z-[40];
   }
   .z-dropdown {
     @apply z-[9];
@@ -355,4 +352,36 @@ body {
 
 .dialog-scrollable-content::-webkit-scrollbar-corner {
   background-color: transparent;
+}
+
+/* Swiper Custom Styles */
+.high-rating-swiper .swiper-slide {
+  height: auto;
+}
+
+.swiper-button-prev-custom,
+.swiper-button-next-custom {
+  align-items: center;
+  justify-content: center;
+  border-radius: 100%;
+  border: 1px solid var(--gray-300);
+  background-color: white;
+  color: var(--gray-600);
+  transition-property: color, border-color;
+  transition-duration: 200ms;
+  transition-timing-function: ease-in-out;
+  cursor: pointer;
+}
+
+.swiper-button-next-custom:hover,
+.swiper-button-prev-custom:hover {
+  border: 1px solid var(--gray-400);
+  color: var(--gray-700);
+}
+
+.swiper-button-prev-custom:disabled,
+.swiper-button-next-custom:disabled {
+  cursor: not-allowed;
+  border: 1px solid var(--gray-200);
+  color: var(--gray-300);
 }
